requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.2.0
django-csp>=4.0
pytz>=2023.3
gunicorn
whitenoise
feedparser
Pillow
